app/requirements.txt
langchain-aws
langchain
python-dotenv
boto3
awscli